Warn user and execute package flash after approval
I believe is the proper way to warn the user.
It does not block the UI thread. Instead the button click
creates its own thread to determine if the pacakges are the same
and to flash the pacakge. The ability to check the packages should
not take  long to process.